+++ lib86cpu/core/emitter/x64/jit.h
@@ -245,17 +245,17 @@ private:
 	void ld_of(x86::Gp dst, x86::Gp aux);
 	void ld_sf(x86::Gp res_dst, x86::Gp aux);
 	void ld_pf(x86::Gp dst, x86::Gp res, x86::Gp aux);
-	void load_mem(uint8_t size);
+	void load_mem(uint64_t size);
 	template<bool dont_write = false>
-	void store_mem(uint8_t size);
+	void store_mem(uint64_t size);
 #ifdef XBOX_CPU
-	void load_ipt(uint8_t size);
-	void load_moffset(uint8_t size, op_info info);
-	void store_ipt(uint8_t size);
-	void store_moffset(uint8_t size, op_info info);
+	void load_ipt(uint64_t size);
+	void load_moffset(uint64_t size, op_info info);
+	void store_ipt(uint64_t size);
+	void store_moffset(uint64_t size, op_info info);
 #endif
-	void load_io(uint8_t size_mode);
-	void store_io(uint8_t size_mode);
+	void load_io(uint64_t size_mode);
+	void store_io(uint64_t size_mode);
 	template<typename T>
 	bool gen_check_io_priv(T port);
 	Label rep_start(Label end);
